+++ configs/neon/compute_wrapper/shell/compute.sh
@@ -1,0 +1,72 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+generate_id() {
+  local -n resvar=${1}
+  printf -v resvar '%08x%08x%08x%08x' ${SRANDOM} ${SRANDOM} ${SRANDOM} ${SRANDOM}
+}
+
+PG_VERSION=${PG_VERSION:-16}
+
+readonly CONFIG_FILE_ORG=/var/db/postgres/configs/config.json
+readonly CONFIG_FILE=/tmp/config.json
+
+first_path="$(ldconfig --verbose 2>/dev/null | grep --invert-match ^$'\t' | cut --delimiter=: --fields=1 | head --lines=1)"
+test "${first_path}" = '/usr/local/lib'
+
+echo "Waiting for pageserver to be ready"
+until nc -z pageserver 6400; do
+  sleep 1
+done
+echo "Pageserver is ready"
+
+cp "${CONFIG_FILE_ORG}" "${CONFIG_FILE}"
+
+if [[ -n "${TENANT_ID:-}" && -n "${TIMELINE_ID:-}" ]]; then
+  tenant_id=${TENANT_ID}
+  timeline_id=${TIMELINE_ID}
+else
+  tenant_id="$(curl -sS -X GET -H "Content-Type: application/json" "http://pageserver:9898/v1/tenant" | jq -r '.[0].id')"
+  if [[ -z "${tenant_id}" || "${tenant_id}" = null ]]; then
+    echo "Creating tenant"
+    generate_id tenant_id
+    curl -sS -X PUT \
+      -H "Content-Type: application/json" \
+      -d '{"mode":"AttachedSingle","generation":1,"tenant_conf":{}}' \
+      "http://pageserver:9898/v1/tenant/${tenant_id}/location_config" >/dev/null
+  fi
+
+  timeline_id=""
+  if [[ "${RUN_PARALLEL:-false}" != "true" ]]; then
+    timeline_id="$(curl -sS -X GET -H "Content-Type: application/json" "http://pageserver:9898/v1/tenant/${tenant_id}/timeline" | jq -r '.[0].timeline_id')"
+  fi
+
+  if [[ -z "${timeline_id}" || "${timeline_id}" = null ]]; then
+    echo "Creating timeline"
+    generate_id timeline_id
+    curl -sS -X POST \
+      -H "Content-Type: application/json" \
+      -d "{\"new_timeline_id\":\"${timeline_id}\",\"pg_version\":${PG_VERSION}}" \
+      "http://pageserver:9898/v1/tenant/${tenant_id}/timeline/" >/dev/null
+  fi
+fi
+
+if [[ ${PG_VERSION} -ge 17 ]]; then
+  ulid_extension=pgx_ulid
+else
+  ulid_extension=ulid
+fi
+
+shared_libraries=$(jq -r '.spec.cluster.settings[] | select(.name=="shared_preload_libraries").value' "${CONFIG_FILE}")
+sed -i "s|${shared_libraries}|${shared_libraries},${ulid_extension}|" "${CONFIG_FILE}"
+sed -i "s|TENANT_ID|${tenant_id}|" "${CONFIG_FILE}"
+sed -i "s|TIMELINE_ID|${timeline_id}|" "${CONFIG_FILE}"
+
+echo "Starting compute node"
+exec /usr/local/bin/compute_ctl \
+  --pgdata /var/db/postgres/compute \
+  -C "postgresql://cloud_admin@localhost:55433/postgres" \
+  -b /usr/local/bin/postgres \
+  --compute-id "compute-${RANDOM}" \
+  --config "${CONFIG_FILE}" \
+  --dev